+++ cmd/client/main.go
@@ -5,227 +5,13 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/cbodonnell/flywheel/client/flow"
-	"github.com/cbodonnell/flywheel/client/input"
+	clientgame "github.com/cbodonnell/flywheel/client/game"
 	"github.com/cbodonnell/flywheel/client/network"
-	"github.com/cbodonnell/flywheel/client/scenes"
 	"github.com/cbodonnell/flywheel/pkg/log"
 	"github.com/cbodonnell/flywheel/pkg/queue"
 	"github.com/cbodonnell/flywheel/pkg/version"
 	"github.com/hajimehoshi/ebiten/v2"
-	"github.com/hajimehoshi/ebiten/v2/ebitenutil"
 )
-
-// Game implements ebiten.Game interface, which has Update, Draw and Layout methods.
-// TODO: move some of this into scenes
-type Game struct {
-	// debug is a boolean value indicating whether debug mode is enabled.
-	debug bool
-	// networkManager is the network manager.
-	networkManager *network.NetworkManager
-	// mode is the current game mode.
-	mode flow.GameMode
-
-	scene scenes.Scene
-}
-
-type NewGameOptions struct {
-	Debug          bool
-	NetworkManager *network.NetworkManager
-}
-
-func NewGame(opts NewGameOptions) (ebiten.Game, error) {
-	g := &Game{
-		debug:          opts.Debug,
-		networkManager: opts.NetworkManager,
-	}
-
-	if err := g.loadMenu(); err != nil {
-		return nil, fmt.Errorf("failed to load menu scene: %v", err)
-	}
-
-	return g, nil
-}
-
-func (g *Game) SetScene(scene scenes.Scene) error {
-	if g.scene != nil {
-		if err := g.scene.Destroy(); err != nil {
-			return fmt.Errorf("failed to destroy previous scene: %v", err)
-		}
-	}
-
-	g.scene = scene
-	if err := g.scene.Init(); err != nil {
-		return fmt.Errorf("failed to initialize scene: %v", err)
-	}
-
-	return nil
-}
-
-func (g *Game) loadMenu() error {
-	menu, err := scenes.NewMenuScene()
-	if err != nil {
-		return fmt.Errorf("failed to create menu scene: %v", err)
-	}
-	if err := g.SetScene(menu); err != nil {
-		return fmt.Errorf("failed to set menu scene: %v", err)
-	}
-	g.mode = flow.GameModeMenu
-	return nil
-}
-
-func (g *Game) loadGame() error {
-	if err := g.networkManager.Start(); err != nil {
-		log.Error("Failed to start network manager: %v", err)
-		g.networkManager.Stop()
-		if err := g.loadNetworkError(); err != nil {
-			return fmt.Errorf("failed to load network error scene: %v", err)
-		}
-		return nil
-	}
-
-	gameScene, err := scenes.NewGameScene(g.networkManager)
-	if err != nil {
-		return fmt.Errorf("failed to create game scene: %v", err)
-	}
-	if err := g.SetScene(gameScene); err != nil {
-		return fmt.Errorf("failed to set game scene: %v", err)
-	}
-	g.mode = flow.GameModePlay
-	return nil
-}
-
-func (g *Game) loadGameOver() error {
-	g.networkManager.Stop()
-
-	gameOver, err := scenes.NewGameOverScene()
-	if err != nil {
-		return fmt.Errorf("failed to create game over scene: %v", err)
-	}
-	if err := g.SetScene(gameOver); err != nil {
-		return fmt.Errorf("failed to set game over scene: %v", err)
-	}
-	g.mode = flow.GameModeOver
-	return nil
-}
-
-func (g *Game) loadNetworkError() error {
-	g.networkManager.Stop()
-
-	networkError, err := scenes.NewErrorScene("Network Error")
-	if err != nil {
-		return fmt.Errorf("failed to create network error scene: %v", err)
-	}
-	if err := g.SetScene(networkError); err != nil {
-		return fmt.Errorf("failed to set network error scene: %v", err)
-	}
-	g.mode = flow.GameModeNetworkError
-	return nil
-}
-
-func (g *Game) Update() error {
-	// Update the network manager
-	if err := g.networkManagerUpdate(); err != nil {
-		return fmt.Errorf("failed to update network manager: %v", err)
-	}
-
-	// Handle input
-	if err := g.handleInput(); err != nil {
-		return fmt.Errorf("failed to handle input: %v", err)
-	}
-
-	// Update the current scene
-	if err := g.scene.Update(); err != nil {
-		return fmt.Errorf("failed to update scene: %v", err)
-	}
-
-	return nil
-}
-
-func (g *Game) networkManagerUpdate() error {
-	if !g.networkManager.IsConnected() {
-		return nil
-	}
-
-	g.networkManager.UpdateServerTime(1.0 / float64(ebiten.TPS()))
-
-	if err := g.checkNetworkManagerErrors(); err != nil {
-		log.Error("Network manager error: %v", err)
-		if err := g.loadNetworkError(); err != nil {
-			return fmt.Errorf("failed to load network error scene: %v", err)
-		}
-	}
-
-	return nil
-}
-
-// checkNetworkManagerErrors checks the network manager for errors and returns any that are found.
-func (g *Game) checkNetworkManagerErrors() error {
-	select {
-	case err := <-g.networkManager.TCPClientErrChan():
-		return fmt.Errorf("TCP client error: %v", err)
-	case err := <-g.networkManager.UDPClientErrChan():
-		return fmt.Errorf("UDP client error: %v", err)
-	default:
-		return nil
-	}
-}
-
-func (g *Game) handleInput() error {
-	switch g.mode {
-	case flow.GameModeMenu:
-		if input.IsPositiveJustPressed() {
-			if err := g.loadGame(); err != nil {
-				return fmt.Errorf("failed to load game scene: %v", err)
-			}
-		}
-	case flow.GameModePlay:
-		if input.IsNegativeJustPressed() {
-			if err := g.loadGameOver(); err != nil {
-				return fmt.Errorf("failed to load game over scene: %v", err)
-			}
-			break
-		}
-	case flow.GameModeOver:
-		if input.IsPositiveJustPressed() {
-			if err := g.loadMenu(); err != nil {
-				return fmt.Errorf("failed to load menu scene: %v", err)
-			}
-		}
-	case flow.GameModeNetworkError:
-		if input.IsPositiveJustPressed() {
-			if err := g.loadMenu(); err != nil {
-				return fmt.Errorf("failed to load menu scene: %v", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (g *Game) Draw(screen *ebiten.Image) {
-	g.scene.Draw(screen)
-	if g.debug {
-		g.drawDebugOverlay(screen)
-	}
-}
-
-func (g *Game) drawDebugOverlay(screen *ebiten.Image) {
-	serverTime, ping := g.networkManager.ServerTime()
-	ebitenutil.DebugPrint(screen, fmt.Sprintf("\n   FPS: %0.1f", ebiten.ActualFPS()))
-	ebitenutil.DebugPrint(screen, fmt.Sprintf("\n\n   TPS: %0.1f", ebiten.ActualTPS()))
-	ebitenutil.DebugPrint(screen, fmt.Sprintf("\n\n\n   Ping: %0.1f", ping))
-	ebitenutil.DebugPrint(screen, fmt.Sprintf("\n\n\n\n   Time: %0.0f", serverTime))
-}
-
-const (
-	DefaultScreenWidth  = 640
-	DefaultScreenHeight = 480
-)
-
-func (g *Game) Layout(outsideWidth, outsideHeight int) (screenWidth, screenHeight int) {
-	return DefaultScreenWidth, DefaultScreenHeight
-}
 
 func main() {
 	debug := flag.Bool("debug", false, "Debug mode")
@@ -249,7 +35,7 @@ func main() {
 		panic(fmt.Sprintf("Failed to create network manager: %v", err))
 	}
 
-	game, err := NewGame(NewGameOptions{
+	game, err := clientgame.NewGame(clientgame.NewGameOptions{
 		Debug:          *debug,
 		NetworkManager: networkManager,
 	})
@@ -257,7 +43,7 @@ func main() {
 		panic(fmt.Sprintf("Failed to create game: %v", err))
 	}
 
-	ebiten.SetWindowSize(DefaultScreenWidth, DefaultScreenHeight)
+	ebiten.SetWindowSize(clientgame.DefaultScreenWidth, clientgame.DefaultScreenHeight)
 	ebiten.SetWindowTitle("Flywheel Client")
 	if err := ebiten.RunGame(game); err != nil {
 		panic(fmt.Sprintf("Failed to run game: %v", err))
